+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- Improve logic in Apple JWS receipts validation.
+### Fixed
+- Fix In-App Purchase validation persist parameter handling for Apple JWS receipts.
 
 ## [3.38.0] - 2026-03-20
 ### Added